+++ tools/tool-base/src/main/java/io/spine/code/gen/java/EntityQueryFactory.java
@@ -30,10 +30,10 @@ import com.google.common.collect.ImmutableList;
 import com.google.errorprone.annotations.Immutable;
 import io.spine.code.gen.java.query.EntityQueryBuilderSpec;
 import io.spine.code.gen.java.query.EntityQuerySpec;
-import io.spine.tools.protoc.method.GeneratedMethod;
-import io.spine.tools.protoc.method.MethodFactory;
-import io.spine.tools.protoc.nested.GeneratedNestedClass;
-import io.spine.tools.protoc.nested.NestedClassFactory;
+import io.spine.tools.protoc.plugin.method.GeneratedMethod;
+import io.spine.tools.protoc.plugin.method.MethodFactory;
+import io.spine.tools.protoc.plugin.nested.GeneratedNestedClass;
+import io.spine.tools.protoc.plugin.nested.NestedClassFactory;
 import io.spine.type.MessageType;
 
 import java.util.List;
